% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{set_client_id}
\alias{set_client_id}
\alias{print_client_id}
\title{Set papago client id}
\usage{
set_client_id(client_id)

print_client_id()
}
\arguments{
\item{client_id}{A string specifying papago client id}
}
\value{
No return value, called to set client id
}
\description{
Save papago client id for the current session. To set it permanently, please add
the following line to your .Renvrion file:

PAPAGO_CLIENT_ID = "YOUR CLIENT ID"
}
