% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{set_client_secret}
\alias{set_client_secret}
\alias{print_client_secret}
\title{Set papago client secret}
\usage{
set_client_secret(client_secret)

print_client_secret()
}
\arguments{
\item{client_secret}{A string specifying papago client secret}
}
\value{
No return value, called to set client secret
}
\description{
Save papago client secret for the current session. To set it permanently, please add
the following line to your .Renvrion file:

PAPAGO_CLIENT_SECRET = "YOUR CLIENT SECRET"
}
