% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Papago text translation}
\usage{
translate(text, source = "en", target = "ko", platform = FALSE)
}
\arguments{
\item{text}{A string or vector text you want to translate}

\item{source}{A string specifying a source language like "en", "ko", "ja", ...}

\item{target}{A string specifying a target language like "en", "ko", "ja", ...}

\item{platform}{FALSE using naver cloud platform or not}
}
\value{
A translated text
}
\description{
\preformatted{
## Examples
translate(text = "Hello, world", source = "en", target = "ko")
translate_data(data = _some_data_frame_, source = "ko", target = "en")
translate_unique_data(data = _some_data_frame_, source = "ja", target = "ko")
}
}
