% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_unique_data}
\alias{translate_unique_data}
\title{Papago unique text translation for data}
\usage{
translate_unique_data(
  data,
  columns,
  source = "en",
  target = "ko",
  platform = FALSE
)
}
\arguments{
\item{data}{A data.frame contains columns you want to translate}

\item{columns}{Columns you want to translate}

\item{source}{A string specifying a source language like "en", "ko", "ja", ...}

\item{target}{A string specifying a target language like "en", "ko", "ja", ...}

\item{platform}{FALSE using naver cloud platform or not}
}
\value{
A data.frame with unique translated columns (contains both input and output columns)
}
\description{
\preformatted{
## Examples
translate_unique_data(data = _some_data_frame_, source = "ja", target = "ko")
}
}
